\name{plot.BMLGrid}
\alias{plot.BMLGrid}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Plot grid
}
\description{
Plot the matrix grid
}
\usage{
plot.BMLGrid(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
S3 Class Grid
}
  \item{\dots}{
Any arguements a general \code{plot()} function take, 
such as \code{xlim(), main()}
}
}
\details{
It can be used as general \code{plot()} function. 
}

\author{
Yuki Liu
}
